pyproject.toml
[project]
name = "zeconfig"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "tomli>=2.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.7",
    "taskipy>=1.14.1",
]

[tool.ruff]
line-length = 79
extend-exclude = ['how_to_use', 'src/core.py', 'src/models.py', 'config']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
